Anti-Spoof AI attendance system/Anti-Spoof AI attendance system/main_project/requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
ultralytics>=8.0.0

# Firebase
firebase-admin>=6.2.0
python-dotenv>=1.0.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0

# Data Handling
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Interface
streamlit>=1.28.0
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
huggingface-hub>=0.17.0
tqdm>=4.66.0
